Here is the prototype for Ship movement I promised.
We now have a very Dadaistic ship bobbing up and down. At 5 seconds we change the amplitude of the waveform describing the ships movement, and at 10 we change the frequency. These operations are SMOOTH! meaning the ship does not suddenly jump position (I hope). 

Next up is to wire in user controls so we an move this ship around ourselves.

Also, if anyone wants to be awesome, adding sprite support would make life easier for when I/someone else wants to play with the course "Preview" option, where we project the ships course across the screen dynamically.

-G