Changed the way we handle keyboard input from a button press system to a polling system (makes more sense to em this way).
Sorry Bjorn, just clobbered most of your changes, but I like this way a bit better :)

As for the bug in AdjustAmplitude, ive been working on it quite a bit, and I know what the bug is, i'm still working on a solution.

The problem is that ArcSin allways returns a result in quadrant 0 or 3 (between -pi/2 and pi/2). Well, ONE, our code only likes values between 0 and 2pi, and second there exist valid solutions in quadrants 1 and 2, and we need to be able to pick the correct solution for the quadrant we are currently in.

If no-one else wants to fix this, i'll take a stab after lunch.