Visualization for the path.
Later I will re-write the path code so that we have a fixed set of points that is cached instead of needing to build a new tuple array each time. But for now, we have a nice visualization going and the first art assets for Jeff to iterate on.

Like my last chekin, I believe this will cause test case problems. I'll need to look more into the testing framework tomorrow, but this is all for tonight.

Enjoy the pretty pictures!