Here are the Title Screen and the Level Select screen.
Yes, the title screen is one huge visual pun. Pat yourself on the back if you get it!

The Levels in the level select are the ones I'm shooting for completing:
-Forest
-Cityscape
-Space Station
-Circuit Board

The blank spots are where the SelectionArrow goes. It should be either blue or red and then flash the other color when a choice is made (I leave this is the programmerses' capable hands).

Also, smaller scaled selection arrows can be used on the two choices on the Title Screen.

Note: I can make widescreen versions on request!