Here are the miscellaneous graphics for the levels.
The Ship graphic has been redrawn and cleaned up as well as embiggened to 128x128!

The selection arrows should now have transparent BGs.

The debris should pop out when you collide with the level and spin wildly and explosively where possible. 

The collectibles may not be used, but, hopefully they can be stuck into the levels to give you a reason to try and get around the more harrowing obstacles with fancy moves rather than just lower the frequency and abuse the pause feature.