Sound Effects!
The Collisions are for hitting the level.
Death is when you die!
The explosions are for enemy ships.
Level_Intro are the names of the levels when you pick them.
Level_Storyline is the story that will play over the level splash screen before the level starts.
Menu_Browse is for choosing options.
Menu_select is for "clicking" them.
Missile Fire is obvious.
Missile Lock 1 2 and 3 are the lockon noises.
Poweruppickup is self-explanatory.
Shield Max is for when your shield gauge reaches 100%.