The (hopefully) last art assets.
Enemy1-4 are different enemy ships.
Explosion1-4 is a 4 frame explosion animation.
Level_Story is the JPG splash screen before starting a level but after picking it from the level select.
LockOn (keep in mind the capital O) is the 3 frame lock on animation.
Missile1-3 is the 3 frame missile animation.
ShipMaxShield displays over the ship when the shield is full (the MaxShield sound also plays).