CHECK IT! Bjorn and I were discussing an alternative game idea where you had a platformer, and nine SECONDARY windows that you can position over the playing field that interact with it in some way. Well, HERE WE GO! Proof of sub-windows!
It works, it's a little laggy, and I need to re-write the on_draw routines to have some semblance of culling (currently the sub windows just overdraws everything... suck! The good news is that we have all the data we need allready for intelligent culling thanks to our spatial trees!